Apparently working pattern for pch with ninja (uses object_outputs and object_depends) source_file_properties

+++ lib_pch/src/pch_lib.cpp
@@ -1,4 +1,4 @@
-#include "pch_header.h"
+// #include "pch_header.h"
 
 #include "..\include\pch_lib.h"
 
@@ -6,5 +6,5 @@ int get_num_lib_pch(void)
 {
     std::vector<std::string> sv = {"hello", "pch", "world"};
 
-    return sv.size();
+    return static_cast<int>(sv.size());
 }
